AR-34 Added new project logic in ProjectComponent
